(emu.h) sound: don't apply gain if value didn't change, also use u32 instead of int for io_connection index

+++ src/devices/machine/k054321.cpp
@@ -127,8 +127,8 @@ void k054321_device::dummy_w(u8 data)
 
 void k054321_device::propagate_volume()
 {
-	double vol = pow(2, (m_volume - 40)/10.0);
+	float vol = powf(2.0f, (m_volume - 40)/10.0f);
 
-	m_speaker->set_input_gain(0, m_active & 2 ? vol * m_left_gain : 0.0);
-	m_speaker->set_input_gain(1, m_active & 1 ? vol * m_right_gain : 0.0);
+	m_speaker->set_input_gain(0, m_active & 2 ? vol * m_left_gain : 0.0f);
+	m_speaker->set_input_gain(1, m_active & 1 ? vol * m_right_gain : 0.0f);
 }
